Split Home page into own file as per KeyPod template #16
